Add predecessors functionality
A Predecessors function is necessary in some DAG algorithms to determine
the "parent" nodes of a task, such as for insetion and replacement of
nodes, as needed by the dagobah project. This algorithm runs in O(n)
time, by simply traversing the list of keys in the graph object, which
is a mapping of nodes to child nodes.