Add all_leaves() functionality
all_leaves() is a function that returns a list of all nodes that do not
have a downstream node. Example use case: In dagobah's future job-in-job
functionality, in order to expand a job reference, we need to make edgs
from all leaves of the inner job to all the downstreams of the job
reference.